Add Integration Tests for IngredientController
- updated BaseintegrationTests methods, added ToQueryString method
- added ToLower() to existence checks in IngredientController and UserController
- added check to see if ingredient is referenced by any recipe in IngredientController
- updated database to force unique ingredient names
- implemented integration tests for IngredientController, covering all CRUD operations

+++ server/RecipesApi/RecipesApi/Controllers/IngredientController.cs
@@ -59,6 +59,13 @@ namespace RecipesApi.Controllers
         [Authorize]
         public async Task<ActionResult<IngredientDTO>> CreateIngredient(CreateIngredientDTO createIngredientDTO)
         {
+            var ingredientExists = await _context.Ingredients.AnyAsync(i => i.Name.ToLower() == createIngredientDTO.Name.ToLower());
+
+            if (ingredientExists)
+            {
+                return BadRequest("Składnik o podanej nazwie już istnieje.");
+            }
+
             // Utwórz nowy składnik na podstawie danych z DTO
             var ingredient = new Ingredient
             {
@@ -112,6 +119,12 @@ namespace RecipesApi.Controllers
             {
                 return NotFound();
             }
+
+            // Sprawdź czy składnik nie jest powiązany z przepisem
+            var isReferenced = await _context.RecipeIngredients
+                .AnyAsync(ri => ri.IngredientId == id);
+            
+            if (isReferenced) return Conflict("Nie można usunąć składnika, ponieważ jest wykorzystywany");
 
             // Usuń składnik
             _context.Ingredients.Remove(ingredient);
